AI_pdf_Reader/requirement.txt
# Core Requirements for PDF Embedding and Semantic Search

# Environment and API Management
python-dotenv>=1.0.0

# OpenAI and LangChain Integrations
langchain>=0.1.0
langchain-openai>=0.0.2

# PDF Processing
PyPDF2>=3.0.0

# Numerical Computing
numpy>=1.22.0

# Optional but Recommended
faiss-cpu>=1.7.0
tiktoken>=0.5.0

# For advanced text splitting (optional)
unstructured>=0.10.0

# Potential alternatives (commented out)
# sentence-transformers
# pypdf